Remove component and rename component2 to component
