+++ crypto/stark/src/component/empty.rs
