+++ crypto/stark/src/component/horizontal.rs
