+++ crypto/stark/src/component/mapped.rs
