+++ crypto/stark/src/component/vertical.rs
